LangGraph/pyproject.toml
[project]
name = "learn-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grandalf>=0.8",
    "ipykernel>=7.1.0",
    "langchain>=1.0.5",
    "langchain-google-genai>=3.0.1",
    "langchain-google-vertexai>=3.0.2",
    "langgraph>=1.0.2",
    "python-dotenv>=1.2.1",
]
